added auth and proxy blueprints;
