+++ .flaskenv
@@ -1,2 +1,6 @@
 FLASK_APP=app
 FLASK_DEBUG=1
+
+CELERY_HOST='redis'
+CELERY_PORT=6379
+CELERY_DB=0
