+++ app/auth/routes.py
@@ -1,0 +1,36 @@
+from flask import render_template, flash, redirect, url_for, request
+from flask_login import login_user, current_user, logout_user
+from werkzeug.urls import url_parse
+
+from . import auth
+
+from .forms import LoginForm
+from ..models import User
+
+
+@auth.route('/login', methods=['GET', 'POST'])
+def login():
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+
+        if user is None or not user.check_password(form.password.data):
+            flash('Invalid username or password')
+            return redirect(url_for('login'))
+
+        login_user(user, remember=form.remember_me.data)
+        next_page = request.args.get('next')
+
+        if not next_page or url_parse(next_page).netloc != '':
+            next_page = url_for('index')
+        return redirect(next_page)
+    return render_template('auth/login.html', title='Sign In', form=form)
+
+
+@auth.route('/logout')
+def logout():
+    logout_user()
+    return redirect(url_for('index'))
