+++ app/proxy/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+proxy = Blueprint('proxy', __name__, template_folder='proxy')
+
+from . import routes
