+++ app/routes.py
@@ -1,55 +1,9 @@
-from flask import render_template, flash, redirect, url_for, request
-from flask_login import login_user, current_user, logout_user, login_required
-from werkzeug.urls import url_parse
+from flask import redirect, url_for
 
-from app import app
-from app.forms import LoginForm, ProxyForm
-from app.models import User, Proxy, get_or_create
+from . import app
 
 
 @app.route('/')
-@app.route('/index')
-# @login_required
 def index():
-    proxies = Proxy.query.all()
-    return render_template('index.html', title='Index', proxies=proxies)
+    return redirect(url_for('proxy.index'))
 
-
-@app.route('/proxy', methods=['GET', 'POST'])
-def proxy():
-    form = ProxyForm()
-    if form.validate_on_submit():
-        host = form.host.data
-        port = form.port.data
-
-        instance, _ = get_or_create(model=Proxy, host=host, port=port)
-        return redirect(url_for('proxy'))
-    return render_template('proxy.html', title='Adding proxy host', form=form)
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-
-        if user is None or not user.check_password(form.password.data):
-            flash('Invalid username or password')
-            return redirect(url_for('login'))
-
-        login_user(user, remember=form.remember_me.data)
-        next_page = request.args.get('next')
-
-        if not next_page or url_parse(next_page).netloc != '':
-            next_page = url_for('index')
-        return redirect(next_page)
-    return render_template('proxy.html', title='Sign In', form=form)
-
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
