+++ app/templates/auth/login.html
@@ -1,0 +1,39 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h1>Sign In</h1>
+    <form action="" method="post" novalidate>
+        {{ form.hidden_tag() }}
+
+        <div class="form-group row">
+            <label for="username" class="col-sm-2 col-form-label">{{ form.username.label }}</label>
+            <div class="col-sm-10">
+                <input type="text" class="form-control" id="username" aria-describedby="usernameHelp" name="username"
+                       placeholder="Enter username." required>
+                <small id="usernameHelp" class="form-text text-muted">Please, enter your username.</small>
+            </div>
+        </div>
+
+        <div class="form-group row">
+            <label for="password" class="col-sm-2 col-form-label">{{ form.password.label }}</label>
+            <div class="col-sm-10">
+                <input type="password" class="form-control" id="password" aria-describedby="passwordHelp"
+                       name="password"
+                       placeholder="Enter password" required>
+                <small id="portHelp" class="form-text text-muted">Please, enter your password.</small>
+            </div>
+        </div>
+        <div class="form-group row">
+            <div class="col-sm-10 offset-sm-2">
+                <div class="form-check">
+                    <input class="form-check-input" type="checkbox" id="remember_me" name="remember_me">
+                    <label class="form-check-label" for="remember_me">
+                        {{ form.remember_me.label }}
+                    </label>
+                </div>
+            </div>
+        </div>
+
+        <button type="submit" class="btn btn-primary">Submit</button>
+    </form>
+{% endblock %}
