+++ app/templates/index.html
@@ -1,29 +1,30 @@
 {% extends "base.html" %}
 
 {% block content %}
-
     <div class="d-flex justify-content-between">
         <div class="p-0">
             <h1>List of proxies</h1>
         </div>
 
-        {# TODO: change#}
-        {% if not current_user.is_authenticated %}
-        <div class="d-flex align-items-center p-0">
-            <a href="{{ url_for('proxy') }}">
-                <button class="btn btn-success">Add Proxy</button>
-            </a>
-        </div>
+        {% if current_user.is_authenticated %}
+            <div class="d-flex align-items-center p-0">
+                <a href="{{ url_for('proxy.add') }}">
+                    <button class="btn btn-success">Add Proxy</button>
+                </a>
+            </div>
         {% endif %}
 
     </div>
     <table class="table table-striped table-bordered">
         <thead>
-            <tr>
-                <th scope="col">#</th>
-                <th scope="col">Host</th>
-                <th scope="col">Port</th>
-            </tr>
+        <tr>
+            <th scope="col">#</th>
+            <th scope="col">Host</th>
+            <th scope="col">Port</th>
+            {% if current_user.is_authenticated %}
+                <th scope="col">Edit</th>
+            {% endif %}
+        </tr>
         </thead>
         <tbody>
         {% for proxy in proxies %}
@@ -31,6 +32,13 @@
                 <th scope="row">{{ loop.index }}</th>
                 <td>{{ proxy.host }} </td>
                 <td>{{ proxy.port }}</td>
+                {% if current_user.is_authenticated %}
+                    <th scope="col">
+                        <a href="{{ url_for('proxy.edit', proxy_id=proxy.id) }}">
+                            <span class="fas fa-pencil-alt text-dark"></span>
+                        </a>
+                    </th>
+                {% endif %}
             </tr>
         {% endfor %}
         </tbody>
