+++ app/templates/proxy/add.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends 'base.html' %}
 
 {% block content %}
     <h1>Adding proxy</h1>
