+++ app/templates/proxy/edit.html
@@ -1,0 +1,21 @@
+{% extends 'base.html' %}
+
+{% block content %}
+    <h1>Editing proxy</h1>
+    <form action="" method="post" novalidate>
+        {{ form.hidden_tag() }}
+        <div class="form-group">
+            <label for="host">{{ form.host.label }}</label>
+            <input type="text" class="form-control" id="host" aria-describedby="hostHelp" name="host"
+                   placeholder="Enter host" required value="{{ proxy.host }}">
+            <small id="hostHelp" class="form-text text-muted">Example host: 198.100.10.138</small>
+        </div>
+        <div class="form-group">
+            <label for="port">{{ form.port.label }}</label>
+            <input type="text" class="form-control" id="port"  aria-describedby="portHelp" name="port"
+                   placeholder="Enter port" required value="{{ proxy.port }}">
+            <small id="portHelp" class="form-text text-muted">Example port: 8080</small>
+        </div>
+        <button type="submit" class="btn btn-primary">Submit</button>
+    </form>
+{% endblock %}
